refactor: move address conversion tool to a single folder
