+++ docusaurus/website/src/components/address-conversion/AddressToScript.tsx
@@ -3,7 +3,7 @@ import { Button, Form, Input, Typography } from "antd";
 import { useFormik } from "formik";
 import "antd/dist/antd.css";
 import styled from "styled-components";
-import {hasShortId, toConfigWithoutShortId} from "../helpers/configHelper";
+import {hasShortId, toConfigWithoutShortId} from "../../helpers/configHelper";
 import { Address, config, helpers, Script } from "@ckb-lumos/lumos";
 type AddressType = "Mainnet" | "Testnet";
