+++ docusaurus/website/src/pages/tools/address-conversion.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Address } from "../../components/Address";
+import { Address } from "../../components/address-conversion/Address";
 import { Sidebar } from "../../components/Sidebar";
 import Layout from "@theme/Layout";
